Workdays scheduler
Added the scheduler page, this page allows you to generate a work schedule based on time period and users wishes.
Added the schedule page, this page is used to view the generated work schedules and edit/delete the work days if something isnt quite right.

+++ VagtPlan.ApiService/Controllers/DepartmentController.cs
@@ -25,7 +25,7 @@ namespace ApiService.Controllers
         }
 
         [HttpGet("get/{id}")]
-        public async Task<ActionResult<Department>> GetDepartment([FromRoute] long id)
+        public async Task<ActionResult<Department>> GetDepartment([FromRoute] int id)
         {
             var department = await _context.Departments.FindAsync(id);
 
@@ -52,7 +52,7 @@ namespace ApiService.Controllers
         }
 
         [HttpPut("edit/{id}")]
-        public async Task<IActionResult> EditDepartment([FromRoute] long id, [FromBody] DepartmentDTO departmentDTO)
+        public async Task<IActionResult> EditDepartment([FromRoute] int id, [FromBody] DepartmentDTO departmentDTO)
         {
             var department = await _context.Departments.FindAsync(id);
             if (department == null)
@@ -80,7 +80,7 @@ namespace ApiService.Controllers
         }
 
         [HttpDelete("delete/{id}")]
-        public async Task<IActionResult> DeleteDepartment([FromRoute] long id)
+        public async Task<IActionResult> DeleteDepartment([FromRoute] int id)
         {
             var department = await _context.Departments.FindAsync(id);
             if (department == null)
@@ -99,7 +99,7 @@ namespace ApiService.Controllers
             return NoContent();
         }
 
-        private bool DepartmentExists(long id)
+        private bool DepartmentExists(int id)
         {
             return _context.Departments.Any(e => e.Id == id);
         }
